Revert "fix: does this load faster"
This reverts commit 3e8dec3c6aa094b205ec7591c055fb4fc8773a59.
